fix: split overlarge windows files and remove plan ids from shipped source
