+++ crates/windows/src/system/window_identity.rs
@@ -4,12 +4,12 @@ use super::process_identity;
 
 /// The immutable identity evidence a resolved window must match.
 ///
-/// KTD3's split: fresh-list verification is strict (title included), while
-/// stored-evidence resolution treats pid + token + app as the immutable
-/// identity and tolerates title drift, logging it as telemetry - a live
-/// window's title legitimately changes (a dirty-marker asterisk, an Electron
-/// target retitling per document), and a hard title check there would fail
-/// drill-down on the very windows 2.4 exists to serve.
+/// Fresh-list verification is strict (title included), while stored-evidence
+/// resolution treats pid + token + app as the immutable identity and tolerates
+/// title drift, logging it as telemetry - a live window's title legitimately
+/// changes (a dirty-marker asterisk, an Electron target retitling per
+/// document), and a hard title check there would fail drill-down on the very
+/// windows this module exists to serve.
 pub(crate) struct WindowIdentityEvidence<'a> {
     pub(crate) handle: super::window_enum::WindowHandle,
     pub(crate) pid: ProcessId,
@@ -64,10 +64,6 @@ impl<'a> WindowIdentityEvidence<'a> {
             .title
             .is_some_and(|expected| live.as_deref() != Some(expected))
         {
-            // Shape only - whether the live title drifted, never the titles
-            // themselves, which are target-derived and would reach any tracing
-            // sink that subscribes (the JSONL trace redacts the `title` key,
-            // but a console/file subscriber would not).
             tracing::debug!("window title changed while immutable source identity remained valid");
         }
         Ok(())
@@ -191,9 +187,6 @@ mod tests {
             return;
         };
         let desktop = unsafe { GetDesktopWindow() };
-        // The strict check compares pid+token+app+title. Using the live image
-        // name as `app` removes the app comparison from the failure, so this
-        // test actually exercises the title arm (and only the title arm).
         let mut win = fake_window(pid, &token, "a-title-that-is-not-the-desktop-title");
         win.app = live_process_app(pid);
         let evidence = WindowIdentityEvidence::from_info(desktop, &win)
